renamed type.h -> box.h
