+++ src/demo_3d.cpp
@@ -4,7 +4,7 @@
 #include "draw/draw.h"
 #include "pcdet-infer-cpu/common/metadata.h"
 #include "pcdet-infer-cpu/common/runtimeconfig.h"
-#include "pcdet-infer-cpu/common/type.h"
+#include "pcdet-infer-cpu/common/box.h"
 #include "pcdet-infer-cpu/pcdet.h"
 #include "pcl.h"
